pyproject.toml
[tool.poetry]
name = "auction"
version = "0.1.0"
description = ""
authors = ["Abdul Magomedov <privetabdul@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.5"
python-dotenv = "^0.19.0"
Pillow = "^8.3.1"
factory-boy = "3.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
